refactor(mobile): merge use-kiloclaw-queries into use-kiloclaw
Eliminates an unnecessary file split that only existed for line count.
All kiloclaw query hooks now live in a single file.
