+++ apps/mobile/src/lib/hooks/use-kiloclaw.ts
@@ -4,13 +4,6 @@ import { resolveContext } from '@/lib/hooks/use-context-query';
 import { useTRPC } from '@/lib/trpc';
 
 export { useKiloClawMutations } from '@/lib/hooks/use-kiloclaw-mutations';
-export {
-  useKiloClawChannelCatalog,
-  useKiloClawConfig,
-  useKiloClawGoogleSetup,
-  useKiloClawSecretCatalog,
-  useStreamChatCredentials,
-} from '@/lib/hooks/use-kiloclaw-queries';
 
 export type InstanceStatus = NonNullable<ReturnType<typeof useKiloClawStatus>['data']>['status'];
 export type GatewayState = NonNullable<
@@ -167,6 +160,96 @@ export function useKiloClawChangelog(organizationId?: string | null) {
     trpc.organizations.kiloclaw.getChangelog.queryOptions(orgInput, {
       enabled: orgEnabled,
       staleTime: 5 * 60_000,
+    })
+  );
+  return isOrg ? org : personal;
+}
+
+export function useKiloClawGoogleSetup(organizationId?: string | null, enabled = true) {
+  const trpc = useTRPC();
+  const { isOrg, personalEnabled, orgEnabled, orgInput } = resolveContext(organizationId, enabled);
+  const personal = useQuery(
+    trpc.kiloclaw.getGoogleSetupCommand.queryOptions(undefined, {
+      enabled: personalEnabled,
+      staleTime: 50 * 60_000,
+    })
+  );
+  const org = useQuery(
+    trpc.organizations.kiloclaw.getGoogleSetupCommand.queryOptions(orgInput, {
+      enabled: orgEnabled,
+      staleTime: 50 * 60_000,
+    })
+  );
+  return isOrg ? org : personal;
+}
+
+export function useKiloClawChannelCatalog(organizationId?: string | null) {
+  const trpc = useTRPC();
+  const { isOrg, personalEnabled, orgEnabled, orgInput } = resolveContext(organizationId);
+  const personal = useQuery(
+    trpc.kiloclaw.getChannelCatalog.queryOptions(undefined, {
+      enabled: personalEnabled,
+      staleTime: 5 * 60_000,
+    })
+  );
+  const org = useQuery(
+    trpc.organizations.kiloclaw.getChannelCatalog.queryOptions(orgInput, {
+      enabled: orgEnabled,
+      staleTime: 5 * 60_000,
+    })
+  );
+  return isOrg ? org : personal;
+}
+
+export function useKiloClawSecretCatalog(organizationId?: string | null) {
+  const trpc = useTRPC();
+  const { isOrg, personalEnabled, orgEnabled, orgInput } = resolveContext(organizationId);
+  const personal = useQuery(
+    trpc.kiloclaw.getSecretCatalog.queryOptions(undefined, {
+      enabled: personalEnabled,
+      staleTime: 5 * 60_000,
+    })
+  );
+  const org = useQuery(
+    trpc.organizations.kiloclaw.getSecretCatalog.queryOptions(orgInput, {
+      enabled: orgEnabled,
+      staleTime: 5 * 60_000,
+    })
+  );
+  return isOrg ? org : personal;
+}
+
+export function useStreamChatCredentials(organizationId?: string | null, enabled = true) {
+  const trpc = useTRPC();
+  const { isOrg, personalEnabled, orgEnabled, orgInput } = resolveContext(organizationId, enabled);
+  const personal = useQuery(
+    trpc.kiloclaw.getStreamChatCredentials.queryOptions(undefined, {
+      enabled: personalEnabled,
+      staleTime: 5 * 60_000,
+    })
+  );
+  const org = useQuery(
+    trpc.organizations.kiloclaw.getStreamChatCredentials.queryOptions(orgInput, {
+      enabled: orgEnabled,
+      staleTime: 5 * 60_000,
+    })
+  );
+  return isOrg ? org : personal;
+}
+
+export function useKiloClawConfig(organizationId?: string | null) {
+  const trpc = useTRPC();
+  const { isOrg, personalEnabled, orgEnabled, orgInput } = resolveContext(organizationId);
+  const personal = useQuery(
+    trpc.kiloclaw.getConfig.queryOptions(undefined, {
+      enabled: personalEnabled,
+      staleTime: 60_000,
+    })
+  );
+  const org = useQuery(
+    trpc.organizations.kiloclaw.getConfig.queryOptions(orgInput, {
+      enabled: orgEnabled,
+      staleTime: 60_000,
     })
   );
   return isOrg ? org : personal;
